requirements.txt
about-time==4.2.1
absl-py==2.3.1
aiohttp==3.9.1
aiosignal==1.4.0
alive-progress==3.1.5
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.12.2
bleach==6.2.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.3.0
executing==2.2.0
fake-useragent==1.4.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.5
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.99.0
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.21.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grapheme==0.6.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
gunicorn==25.3.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaggle==1.7.4.5
kiwisolver==1.4.8
langextract==1.0.5
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mbox-to-json==1.0.5
ml_collections==1.1.0
more-itertools==10.7.0
MouseInfo==0.1.3
multidict==6.6.3
multiprocess==0.70.18
nest-asyncio==1.6.0
numpy==2.3.1
openai==1.90.0
outcome==1.3.0.post0
p2j==1.3.2
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyparsing==3.2.3
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.17
PyQt5_sip==12.17.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
rsa==4.9.1
rubicon-objc==0.5.1
scikit-learn==1.7.0
scikit-llm==1.4.1
scipy==1.16.0
selenium==4.33.0
setuptools==75.3.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
steam==1.4.4
tenacity==8.5.0
text-unidecode==1.3
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.66.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.4.0
valve==0.0.0
vdf==3.4
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.12.0
wsproto==1.2.0
yarl==1.20.1
